Separate the contrast property read function and contrast calculation
This is used in random() to randomize the colors until we find a color
pair with a suitable contrast ratio. However, this emits signals and is
generally wasteful (as it goes through a bunch of layers in Qt & Rust)
when we just need to set it once - at the end.

So now there's a freestanding contrast_for_colors function to grab the
contrast ratio, and contrast simply calls this.